Fix font inconsistencies
The default font families are Lucida Grande, Lucida Sans Unicode, **GNU
Unifont**, Verdana, Helvetica, and the catch-all sans-serif. However, Unifont
is a bitmap monospace font, unlike the rest of them, which are vectorised
proportional fonts. Switch Unifont to Dejavu Sans, which is more appropriate
and consistent in this context.

Additionally, two stylesheets use a different font priority of Lucida Grande,
Lucida Sans Unicode, **Arial Unicode MS**, GNU Unifont, and sans-serif, which
is probably a typo. Switch them to the default fonts, except for the IE-related
stylesheets.